Issue #1019976 by halstead: introduce an entirely new import methodology based on git-fast-export.
This new approach (currently only functioning for parsing brand new
repos) uses the cluster of data produced by git-fast-export to minimize
the number of git calls that need to be made, then bypasses the API
in order to perform massive aggregate inserts. Benchmarks generally
indicate it to be more than 100x faster than the old approach, as well
as more accurate. At one point, it was even closer to 1000x faster...but
that was before it became feature-complete.